+++ lib/fastererer/cli.rb
@@ -2,6 +2,7 @@
 
 require 'optparse'
 require_relative 'file_traverser'
+require_relative 'formatters'
 require_relative 'painter'
 require_relative 'version'
 
@@ -10,15 +11,17 @@ module Fastererer
     OFFENSES_FOUND_STATUS = 1
     USAGE_ERROR_STATUS = 2
 
-    def self.execute
+    def self.execute(out: $stdout, err: $stderr)
       options = parse_options(ARGV.dup)
+      formatter = Formatters.fetch(options[:format] || 'text').new(out:, err:)
+    rescue UnknownFormatError, OptionParser::ParseError => e
+      err.puts(e.message)
+      exit USAGE_ERROR_STATUS
+    else
       Painter.disable! if options[:no_color]
-      file_traverser = Fastererer::FileTraverser.new(options[:path])
+      file_traverser = FileTraverser.new(options[:path], formatter:)
       file_traverser.traverse
       exit_with_status_for(file_traverser)
-    rescue OptionParser::ParseError => e
-      warn e.message
-      exit USAGE_ERROR_STATUS
     end
 
     def self.exit_with_status_for(file_traverser)
@@ -35,6 +38,9 @@ module Fastererer
     def self.build_parser(options)
       OptionParser.new do |opts|
         opts.banner = 'Usage: fastererer [options] [path]'
+        opts.on('-f', '--format FORMAT', 'Output format: text (default), json, rdjsonl') do |format|
+          options[:format] = format
+        end
         opts.on('--no-color', 'Disable ANSI color in output') { options[:no_color] = true }
         opts.on('-h', '--help', 'Show this help') { show_help_and_exit(opts) }
         opts.on('-v', '--version', 'Show version') { show_version_and_exit }
@@ -47,7 +53,7 @@ module Fastererer
     end
 
     def self.show_version_and_exit
-      puts Fastererer::VERSION
+      puts VERSION
       exit
     end
   end